+++ packages/plugin-autocapture-browser/src/autocapture-plugin.ts
@@ -265,7 +265,6 @@ export const autocapturePlugin = (
     // Create subscriptions
     const clickTrackingSubscription = trackClicks({
       allObservables,
-      options: options as AutoCaptureOptionsWithDefaults,
       amplitude,
       shouldTrackEvent: shouldTrackEvent,
       evaluateTriggers: evaluateTriggers.evaluate.bind(evaluateTriggers),
@@ -290,7 +289,9 @@ export const autocapturePlugin = (
       shouldTrackEvent,
       shouldTrackActionClick: shouldTrackActionClick,
     });
-    subscriptions.push(actionClickSubscription);
+    if (actionClickSubscription) {
+      subscriptions.push(actionClickSubscription);
+    }
 
     /* istanbul ignore next */
     config?.loggerProvider?.log(`${name} has been successfully added.`);