+++ packages/plugin-session-replay-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.23.1...@amplitude/plugin-session-replay-browser@1.23.2) (2025-11-05)
+
+**Note:** Version bump only for package @amplitude/plugin-session-replay-browser
+
+
+
+
+
 ## [1.23.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-session-replay-browser@1.23.0...@amplitude/plugin-session-replay-browser@1.23.1) (2025-10-29)
 
 **Note:** Version bump only for package @amplitude/plugin-session-replay-browser